pyproject.toml
[tool.poetry]
name = "curso-django"
version = "0.1.0"
description = ""
authors = ["Sérgio Dantas <sergiodnts828@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
Django = "~4.2.7"
Pillow = "~10.1.0"
python-dotenv = "1.0.0"
pytest-dotenv = "^0.5.2"
djangorestframework = "^3.15.2"

[tool.poetry.group.dev.dependencies]
Faker = "~20.1.0"
pytest = "~7.4.3"
pytest-django = "~4.7.0"
parameterized = "~0.9.0"
coverage = "~7.3.2"
selenium = "~4.16.0"
django-debug-toolbar = "~4.2.0"
ruff = "~0.1.14"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.lint]
ignore = ["F401"]

[tool.ruff.format]
quote-style = "single"
